requirements.txt
# requirements.txt
langchain
langchain-community
langchain-groq
streamlit
plotly
pandas
sqlalchemy
python-dotenv